Set default values when Healthcheck time vars are not set
